+++ src/index.ts
@@ -2,6 +2,7 @@ import Discord from 'discord.js'
 import { Client } from 'discord.js'
 
 const client: Client = new Discord.Client()
+const prefix = "?EXAMPLE"
 
 if (!process.env.DISCORDTOKEN) {
     console.error(
@@ -20,6 +21,15 @@ if (!process.env.DISCORDTOKEN) {
  */
 client.on('ready', () => {
     console.log('Bot ready')
+})
+
+client.on('message', message => {
+	if(!message.content.startsWith(prefix) || message.author.bot) return;
+
+	const args = message.content.slice(prefix.length).split(/ +/);
+
+	console.log(args);
+	message.channel.send(args);
 })
 
 // https://discord.com/developers/applications
